+++ src/Misc.cpp
@@ -43,7 +43,7 @@ namespace misc
 	//dirty hack
 	int FtoIround(float d)
 	{
-		return d<0 ? d - .5f : d + .5f;
+		return (int)(d<0 ? d - .5f : d + .5f);
 	}
 
 	char roundc(float a)
@@ -76,7 +76,7 @@ namespace misc
 
 	int getADTCord(float cord)
 	{
-		return cord / 533.33333f;
+		return (int)(cord / 533.33333f);
 	}
 }
